pyproject.toml
[project]
name = "trialfield-core"
version = "0.0.1"
description = "On-farm trial designer — geometry engine and CLI"
requires-python = ">=3.11"
dependencies = [
    "shapely>=2.0",
    "pyproj>=3.6",
    "pyshp>=2.3",
    "openpyxl>=3.1",
    "matplotlib>=3.8",
    "pydantic>=2.5",
    "anthropic>=0.40",
    "typer>=0.12",
    "pypdf>=4.0",
    "python-docx>=1.1",
    "requests>=2.31",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "ruff>=0.6", "mypy>=1.8"]

[project.scripts]
trialfield = "trialfield_core.cli:app"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["trialfield_core*"]
exclude = ["reference*", "tests*"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
